chore: remove server-side mods from additive
as additive is strictly client-only and is not meant to be compatible with server environments, removes the server-side mods such as c2me, krypton, servercore, and vmp since they are unnecessary to keep in the packwiz project. they will stay in adrenaline of course, as that is compatible with both client and server environments.